ref(nav): move whatsnew outside of the folder
