--- static/app/views/navigation/primary/whatsNew/item.tsx
@@ -1,87 +1,0 @@
-import {useCallback} from 'react';
-import styled from '@emotion/styled';
-
-import {Tag} from '@sentry/scraps/badge';
-import {Stack} from '@sentry/scraps/layout';
-import {ExternalLink} from '@sentry/scraps/link';
-
-import {t} from 'sentry/locale';
-import type {Broadcast} from 'sentry/types/system';
-import {trackAnalytics} from 'sentry/utils/analytics';
-import useOrganization from 'sentry/utils/useOrganization';
-
-export const BROADCAST_CATEGORIES: Record<NonNullable<Broadcast['category']>, string> = {
-  announcement: t('Announcement'),
-  feature: t('New Feature'),
-  blog: t('Blog Post'),
-  event: t('Event'),
-  video: t('Video'),
-};
-
-interface BroadcastPanelItemProps extends Pick<
-  Broadcast,
-  'hasSeen' | 'category' | 'title' | 'message' | 'link' | 'mediaUrl'
-> {}
-
-export function WhatsNewItem({
-  hasSeen,
-  title,
-  message,
-  link,
-  mediaUrl,
-  category,
-}: BroadcastPanelItemProps) {
-  const organization = useOrganization();
-
-  const handlePanelClicked = useCallback(() => {
-    trackAnalytics('whats_new.link_clicked', {organization, title, category});
-  }, [organization, title, category]);
-
-  return (
-    <SidebarPanelItemRoot>
-      <Stack align="start" marginBottom="lg">
-        {category && (
-          <CategoryTag variant="muted">{BROADCAST_CATEGORIES[category]}</CategoryTag>
-        )}
-        <Title hasSeen={hasSeen} href={link} onClick={handlePanelClicked}>
-          {title}
-        </Title>
-        <Message>{message}</Message>
-      </Stack>
-      {mediaUrl && <Media src={mediaUrl} alt={title} />}
-    </SidebarPanelItemRoot>
-  );
-}
-
-const SidebarPanelItemRoot = styled('div')`
-  line-height: 1.5;
-  margin: 0 ${p => p.theme.space['2xl']};
-  padding: ${p => p.theme.space.xl} 0;
-
-  :not(:first-child) {
-    border-top: 1px solid ${p => p.theme.tokens.border.primary};
-  }
-`;
-
-const Title = styled(ExternalLink)<Pick<BroadcastPanelItemProps, 'hasSeen'>>`
-  font-size: ${p => p.theme.font.size.lg};
-  color: ${p => p.theme.tokens.content.accent};
-  ${p => !p.hasSeen && `font-weight: ${p.theme.font.weight.sans.medium}`};
-  &:focus-visible {
-    box-shadow: none;
-  }
-`;
-
-const Message = styled('div')`
-  color: ${p => p.theme.tokens.content.secondary};
-`;
-
-const Media = styled('img')`
-  border-radius: ${p => p.theme.radius.md};
-  border: 1px solid ${p => p.theme.colors.gray200};
-  max-width: 100%;
-`;
-
-const CategoryTag = styled(Tag)`
-  margin-bottom: ${p => p.theme.space.md};
-`;
